Don't use link time optimisation.
Link time optimisation seems to make the solver run a few percent faster, but
I'm turning it off by default to make comparisons with programs that do not use
LTO fairer.  LTO also makes compilation take a little longer, and I've read
that it might be buggy on older compilers?

This is a marginal decision, and it is fine to compile with -flto.